requirements/base.txt
Flask
httpie
python-dotenv
